Reorganize account mutations
