+++ app/graphql/mutations/account/delete_account_mutation.rb
@@ -1,0 +1,28 @@
+# frozen_string_literal: true
+
+module Mutations
+  module Account
+    DeleteAccountMutation = GraphQL::Field.define do
+      type types.Boolean
+      description <<~DESCRIPTION
+        Deletes the account of the currently signed in user.
+        Returns `true` if it was successful and `null` if there was an error.
+      DESCRIPTION
+
+      argument :password, !types.String do
+        description 'Password of the current user to confirm the deletion'
+      end
+
+      resource ->(_root, _arguments, context) { context[:current_user] }
+      resolve DeleteAccountResolver.new
+    end
+
+    # GraphQL mutation to delete the current user account
+    class DeleteAccountResolver
+      def call(user, arguments, _context)
+        user.destroy if user.valid_password?(arguments[:password])
+        true
+      end
+    end
+  end
+end
